auth-services/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Deprecated==1.2.14
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gcloud==0.18.3
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauth2client==4.1.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-box==7.1.1
python-jwt==4.1.0
pytz==2023.3.post1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
urllib3==1.26.18
Werkzeug==3.0.1
wrapt==1.16.0
